// src/types/store/addAcceptanceTaskForm.ts
// export interface addAcceptanceTaskFormData {
// 	title: string;
// 	dateStart: string;
// 	dateEnd: string;
// 	userId: string;
// 	productIds: Array<string>;
// 	startPointIds: Array<string>;
// 	endPointIds: Array<string>;
// }

export interface addAcceptanceTaskFormDataArrays {
	products: Array<string>;
	startPoints: Array<string>;
	endPoints: Array<string>;
}

export interface addAcceptanceTaskFormDataFields {
	title: string;
	dateStart: string;
	dateEnd: string;
	userId: string;
}